requirements.txt
Flask
flask-socketio
requests
beautifulsoup4
lxml
pandas
pyspellchecker
